Keep VS diagnostics requests open until the workspace changes to align with public spec
