+++ src/Features/LanguageServer/Protocol/Handler/Diagnostics/Public/PublicWorkspacePullDiagnosticsHandler.cs
@@ -5,11 +5,8 @@
 using System;
 using System.Collections.Immutable;
 using System.Linq;
-using System.Threading;
-using System.Threading.Tasks;
 using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.Diagnostics;
-using Microsoft.CodeAnalysis.EditAndContinue;
 using Microsoft.CodeAnalysis.Options;
 using Microsoft.VisualStudio.LanguageServer.Protocol;
 using Roslyn.Utilities;
@@ -22,39 +19,15 @@ namespace Microsoft.CodeAnalysis.LanguageServer.Handler.Diagnostics.Public;
 using WorkspaceDiagnosticPartialReport = SumType<WorkspaceDiagnosticReport, WorkspaceDiagnosticReportPartialResult>;
 
 [Method(Methods.WorkspaceDiagnosticName)]
-internal sealed class PublicWorkspacePullDiagnosticsHandler : AbstractPullDiagnosticHandler<WorkspaceDiagnosticParams, WorkspaceDiagnosticPartialReport, WorkspaceDiagnosticReport?>, IDisposable
+internal sealed class PublicWorkspacePullDiagnosticsHandler(
+    LspWorkspaceManager workspaceManager,
+    LspWorkspaceRegistrationService registrationService,
+    IDiagnosticAnalyzerService analyzerService,
+    IDiagnosticsRefresher diagnosticRefresher,
+    IGlobalOptionService globalOptions)
+    : AbstractWorkspacePullDiagnosticsHandler<WorkspaceDiagnosticParams, WorkspaceDiagnosticPartialReport, WorkspaceDiagnosticReport?>(
+        workspaceManager, registrationService, analyzerService, diagnosticRefresher, globalOptions), IDisposable
 {
-    private readonly LspWorkspaceRegistrationService _workspaceRegistrationService;
-    private readonly LspWorkspaceManager _workspaceManager;
-
-    /// <summary>
-    /// Flag that represents whether the LSP view of the world has changed.
-    /// It is totally fine for this to somewhat over-report changes
-    /// as it is an optimization used to delay closing workspace diagnostic requests
-    /// until something has changed.
-    /// </summary>
-    private int _lspChanged = 0;
-
-    public PublicWorkspacePullDiagnosticsHandler(
-        LspWorkspaceManager workspaceManager,
-        LspWorkspaceRegistrationService registrationService,
-        IDiagnosticAnalyzerService analyzerService,
-        IDiagnosticsRefresher diagnosticRefresher,
-        IGlobalOptionService globalOptions)
-        : base(analyzerService, diagnosticRefresher, globalOptions)
-    {
-        _workspaceManager = workspaceManager;
-        _workspaceRegistrationService = registrationService;
-
-        _workspaceRegistrationService.LspSolutionChanged += OnLspSolutionChanged;
-        _workspaceManager.LspTextChanged += OnLspTextChanged;
-    }
-
-    public void Dispose()
-    {
-        _workspaceManager.LspTextChanged -= OnLspTextChanged;
-        _workspaceRegistrationService.LspSolutionChanged -= OnLspSolutionChanged;
-    }
 
     /// <summary>
     /// Public API doesn't support categories (yet).
@@ -123,13 +96,6 @@ internal sealed class PublicWorkspacePullDiagnosticsHandler : AbstractPullDiagno
         };
     }
 
-    protected override ValueTask<ImmutableArray<IDiagnosticSource>> GetOrderedDiagnosticSourcesAsync(
-        WorkspaceDiagnosticParams diagnosticParams, RequestContext context, CancellationToken cancellationToken)
-    {
-        // Task list items are not reported through the public LSP diagnostic API.
-        return WorkspacePullDiagnosticHandler.GetDiagnosticSourcesAsync(context, GlobalOptions, cancellationToken);
-    }
-
     protected override ImmutableArray<PreviousPullResult>? GetPreviousResults(WorkspaceDiagnosticParams diagnosticsParams)
     {
         return diagnosticsParams.PreviousResultId.Select(id => new PreviousPullResult
@@ -142,48 +108,5 @@ internal sealed class PublicWorkspacePullDiagnosticsHandler : AbstractPullDiagno
         }).ToImmutableArray();
     }
 
-    private void OnLspSolutionChanged(object? sender, WorkspaceChangeEventArgs e)
-    {
-        UpdateLspChanged();
-    }
-
-    private void OnLspTextChanged(object? sender, EventArgs e)
-    {
-        UpdateLspChanged();
-    }
-
-    private void UpdateLspChanged()
-    {
-        Interlocked.Exchange(ref _lspChanged, 1);
-    }
-
-    protected override async Task WaitForChangesAsync(RequestContext context, CancellationToken cancellationToken)
-    {
-        // Spin waiting until our LSP change flag has been set.  When the flag is set (meaning LSP has changed),
-        // we reset the flag to false and exit out of the loop allowing the request to close.
-        // The client will automatically trigger a new request as soon as we close it, bringing us up to date on diagnostics.
-        while (Interlocked.CompareExchange(ref _lspChanged, value: 0, comparand: 1) == 0)
-        {
-            // There have been no changes between now and when the last request finished - we will hold the connection open while we poll for changes.
-            await Task.Delay(TimeSpan.FromMilliseconds(100), cancellationToken).ConfigureAwait(false);
-        }
-
-        context.TraceInformation("Closing workspace/diagnostics request");
-        // We've hit a change, so we close the current request to allow the client to open a new one.
-        return;
-    }
-
-    internal TestAccessor GetTestAccessor() => new(this);
-
-    internal readonly struct TestAccessor
-    {
-        private readonly PublicWorkspacePullDiagnosticsHandler _handler;
-
-        public TestAccessor(PublicWorkspacePullDiagnosticsHandler handler)
-        {
-            _handler = handler;
-        }
-
-        public void TriggerConnectionClose() => _handler.UpdateLspChanged();
-    }
+    internal override TestAccessor GetTestAccessor() => new(this);
 }
